me faltaba el ejecutar el purge de los templates dentro del lector.

+++ cubieDevice/src/main/java/ar/com/dcsys/firmware/model/RegenerateFingerprintsInReader.java
@@ -9,9 +9,10 @@ import javax.inject.Inject;
 import javax.inject.Provider;
 
 import ar.com.dcsys.data.fingerprint.FingerprintDAO;
-import ar.com.dcsys.data.log.AttLog;
+import ar.com.dcsys.firmware.cmd.template.ClearAllTemplate.ClearAllTemplateResult;
 import ar.com.dcsys.firmware.database.FingerprintMappingDAO;
 import ar.com.dcsys.firmware.leds.Leds;
+import ar.com.dcsys.firmware.reader.PurgeTemplates;
 import ar.com.dcsys.security.Fingerprint;
 
 public class RegenerateFingerprintsInReader implements Cmd {
@@ -21,17 +22,20 @@ public class RegenerateFingerprintsInReader implements Cmd {
 	
 	private final Leds leds;
 	private final Provider<PersistFingerprint> persistFingerprint;
+	private final Provider<PurgeTemplates> purgeTemplates;
 	private final FingerprintDAO fingerprintDAO;
 	private final FingerprintMappingDAO fingerprintMappingDAO;
 	
 	@Inject
 	public RegenerateFingerprintsInReader(Leds leds,
 						Provider<PersistFingerprint> persistFingerprint,
+						Provider<PurgeTemplates> purgeTemplates,
 						FingerprintDAO fingerprintDAO,
 						FingerprintMappingDAO fingerprintMappingDAO) {
 
 		this.leds = leds;
 		this.persistFingerprint = persistFingerprint;
+		this.purgeTemplates = purgeTemplates;
 		this.fingerprintDAO = fingerprintDAO;
 		this.fingerprintMappingDAO = fingerprintMappingDAO;
 	}
@@ -62,7 +66,7 @@ public class RegenerateFingerprintsInReader implements Cmd {
 	private boolean cancel = false;
 	
 	private class CountContainer { int count = 0; };
-	
+	private class BooleanContainer { boolean v = false; };
 	
 	@Override
 	public void execute() {
@@ -84,9 +88,56 @@ public class RegenerateFingerprintsInReader implements Cmd {
 				}
 			};
 			
+			final BooleanContainer purgeError = new BooleanContainer();
+			
+			PurgeTemplates pt = purgeTemplates.get();
+			pt.execute(new ClearAllTemplateResult() {
+				@Override
+				public void onSuccess(int number) {
+					purgeError.v = false;
+					try {
+						remote.sendText("ok " + String.valueOf(number) + " reader purged");
+					} catch (IOException e) {
+						logger.log(Level.SEVERE,e.getMessage(),e);
+						try {
+							remote.sendText("ERROR " + e.getMessage());
+						} catch (IOException e1) {
+							logger.log(Level.SEVERE,e1.getMessage(),e1);
+						}
+					}
+					
+				}
+				@Override
+				public void onFailure(int errorCode) {
+					purgeError.v = true;
+					try {
+						remote.sendText("ERROR " + String.valueOf(errorCode));
+					} catch (IOException e1) {
+						logger.log(Level.SEVERE,e1.getMessage(),e1);
+					}
+				}
+				@Override
+				public void onCancel() {
+					purgeError.v = true;
+					try {
+						remote.sendText("ERROR canceled");
+					} catch (IOException e1) {
+						logger.log(Level.SEVERE,e1.getMessage(),e1);
+					}
+				}
+			});
+			
+			// si existió un error en el purge de las huellas retorno y no sigo con nada.
+			if (purgeError.v) {
+				return;
+			}
 			
 			fingerprintMappingDAO.deleteAll();
+			remote.sendText("ok mappings deleted");
+			
 			List<String> ids = fingerprintDAO.findAll();
+			remote.sendText("ok " + ids.size());
+			
 			for (String id : ids) {
 				Fingerprint fp = fingerprintDAO.findById(id);
 				PersistFingerprint pf = persistFingerprint.get();
